Create character_props_output.py
this is a trick that is mostly ignored and can replace the need for separate chars for simpler material fading etc 

these are props for edit mode (typing), but even there they don't work quite as expected (as u cannot select some chars and check bold)
So I make it not per character, but as selecting portion of text.( Also so that we don't need to get this inside a loop.)

However, this has some problem. I should move the  *underline size, underline position and smallcaps scale* into the text obj out, cause they are per object, not per char. Only the bool is per character.

ps. Yes, I know, you will absolutely love so many useProp .. :)