this is the right version
Sorry, the variant you merged in master is not the right one. And is not working (as exceptions)

Here's the way, without Direction variants, so for Dir 000 we do nothing
but for the rest, that is the way to make exceptions.


In the last var I have changed the guide exceptions cause they were not working right. 
The def of world vec was irrelevant style thing, the essential was to have != vector ((0, 0, 0)) and not use the corrected guide ...
However, now I have taken out the Dir exceptions and those world vecs...