cleanup sockets/info.py file
It now uses a cache that is filled when all socket classes has been
imported.
This new structure allows faster access to attributes of data types and
id names.
Furthermore it is easy to rebuild the cache when new socket types are
registered during runtime (eg through an extension system in the future)