Remove Update Object Matrices node
This node does not work on many levels. First, the location, rotation,
and scale are computed from the evaluated object, which is actually not
evaluated in the active depsgraph assuming the user altered the object
state, which is usually the case. So this node effectively reset the
matrices to their previous values, undoing the user's work. This can be
worked around by using the matrix basis or the transforms of the
original object, but this brings us to the next problem. Second, the
node does not take into consideration any parenting, constraints,
animations or any other depsgraph relation. So the node is practically
broken and I see no reason to keep it.