Use Object.to_curve method for spline reading
Using this method allow us to evaluate the modifiers on the splines
which can now be controlled through the Apply Modifiers input in the
Splines From Object node. This also allow use to get the splines
representing a text object, which are now allowed as object types for
spline reading.