Optimize execution units setup
Currently, all execution units are setup even if non of them are used by
the main execution units currently executing. This is highly inefficient
for setups where passive node trees are used as template libraries. This
patch optimize this by only setting up units that are references by the
main units currently executing. This yields a 50x speedup in the sample
shared in #1783.

Closes #1808.