Tempo events being retrieved as a list of tempo events from MIDI File Reader, including time in quarter notes as well.
New file midi_tempo_event.py in data_structures, so it can be shared with utils/midi.py and sockets/midi_tempo_event.py. It also has the event time in quarter notes. It already had in ticks and in seconds.

New file midi_tempo_event.py in socktes, so tempos can be output from midi_file_reader as a single tempo event or as a list of tempo events.

midi_file_reader.py has a new output, which can be a tempo event or a list of tempo events.

utils/midi.py return now a list of tempo events. These tempo events contain the event time in ticks, in quarter notes and in seconds and the tempo in microseconds per quarter note (MIDI default).

TODO: create a MIDI Tempo Event Info that gives access to the list of tempos, list of tempo times in seconds and list of tempo times in quarter notes.