Fix error when subprogram is deleted
When a subprogram is deleted while still being referenced by an invoke
node, an error occurs and Animation Nodes stops working. This happens
because the invoke node still contains the identifier of the deleted
subprogram, and the code assumes it is valid. To fix this, check if
there is in fact a subprogram with the given identified before making
any use of it.

Fixes #1893.