Fixed incorrect positioning of node UI extensions
Blender changed the value of its DPI value to hold the actual monitor
DPI, as opposed to additionally dividing by the pixel size for some
reason, that is, the line width. Consequently, we no longer need to
multiply by the pixel size to obtain the original DPI factor. We also
take the opportunity to omit the DPI argument of the BLF size method
since it is no deprecated. Finally, the line width property is now used
instead of the DPI factor where it make sense.

Fixes #1905.