Fix node tree not executing due to active sequence
We use the frame changed application handler to execute node tree when
an animation render is taking place. However, Blender has a limitation
where the frame changed handler doesn't execute if a scene is not being
rendered during the rendering pipeline, for instance, when a sequence
editor is active and doesn't reference a scene sequence strip. So as a
workaround, we use the render pre handler to execute the node tree as a
frame change if we determine that the frame changed handler will not run
due to the aforementioned reasons.

This is not an ideal solution, but a necessary workaround that we have
to implement. Fixing the workaround in Blender itself turned out to be a
bit complex, because the handlers are executed at difference parts of
the code, so double execution can easily happen.

Other similar situations might exist, but we shall solve them in
separate commits when they come up.

Fixes #1902.